Extracted itemSettings and innetElementType
