+++ src/js/shot-generator-new/components/HandPresetsEditor/HandPresetsEditorItem.js
@@ -3,9 +3,9 @@ import path from 'path'
 import { remote } from 'electron'
 import classNames from 'classnames'
 import ThumbnailRenderer from '../../ThumbnailRenderer'
-import { useMemo } from 'react'
+import React, { useMemo } from 'react'
 import { createdMirroredHand, applyChangesToSkeleton, getOppositeHandName } from '../../../utils/handSkeletonUtils'
-import { GUTTER_SIZE, ITEM_WIDTH, ITEM_HEIGHT, IMAGE_HEIGHT, IMAGE_WIDTH } from './ItemSettings'
+import { GUTTER_SIZE, ITEM_WIDTH, ITEM_HEIGHT, IMAGE_HEIGHT, IMAGE_WIDTH } from '../../utils/InspectorElementsSettings'
 import { getSceneObjects } from '../../../shared/reducers/shot-generator'
 import clampInstance from '../../utils/clampInstance'
