--- src/js/shot-generator-new/components/PosePresetsEditor/ItemSettings.js
@@ -1,10 +1,0 @@
-const NUM_COLS = 4
-const GUTTER_SIZE = 5
-const ITEM_WIDTH = 68
-const ITEM_HEIGHT = 132
-
-const IMAGE_WIDTH = ITEM_WIDTH
-const IMAGE_HEIGHT = 100
-const CHARACTER_MODEL = { id: 'adult-male', type: 'character' }
-
-export { NUM_COLS, GUTTER_SIZE, ITEM_WIDTH, ITEM_HEIGHT, IMAGE_HEIGHT, IMAGE_WIDTH, CHARACTER_MODEL }
